requirements.txt
PyQt5
geopandas
PyQtWebEngine
folium
scipy
matplotlib
pandapipes
geopy
overpy
geojson
scikit-learn
PyPDF2
reportlab
numpy_financial